# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Recipe Book",
    "description": "API System for Recipe Book",
    "version": "1.0.0",
    "contact": {
      "name": "CSE341-GROUP08 Project Github Repo Link",
      "url": "https://github.com/chrmux/cse341-group08-project"
    }
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/recipes/{id}": {
      "get": {
        "tags": [
          "Recipes"
        ],
        "summary": "Fetch item by Id",
        "description": "Find one from the database",
        "operationId": "getRecipeById",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "Id of the recipe"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "put": {
        "tags": [
          "Recipes"
        ],
        "summary": "Update Recipe Data",
        "description": "Update the existing Recipe in database",
        "operationId": "updateRecipe",
        "parameters": [{
            "name": "id",
            "in": "path",
            "required": "true",
            "description": "id of the recipe",
            "type": "string"
          },
          {
            "name": "recipes",
            "description": "Recipe params to update the recipe",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "photo": {
                  "type": "string",
                  "example": "any"
                },
                "title": {
                  "type": "string",
                  "example": "any"
                },
                "ingredients": {
                  "type": "string",
                  "example": "any"
                },
                "instruction": {
                  "type": "string",
                  "example": "any"
                },
                "creator": {
                  "type": "string",
                  "id": {
                    "type": "string"
                  }
                },
                "keywords": {
                  "type": "string"
                },
                "culture": {
                  "type": "string"
                },
                "level": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "Recipes"
        ],
        "summary": "Delete Item",
        "description": "Delete Recipe from the database",
        "operationId": "deleteRecipe",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "id",
          "in": "path",
          "required": "true",
          "description": "Id of the recipe",
          "type": "string"
        }],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
...
}